feat: extract token detail table columns
